So I did a few things. 1) I substituted the new ways of opening camera OR camera roll for the old way. Located in the same folder, these views include CameraButtonView which is a view for an icon, ImagePicker which is a view that also existed in the previous thing that you included, but the one that I included is different and also some funky code that I understand a bit and would love to explain, and finally, UsingCamera which is never actually used in the app but I use as a template of how to use the new code. 2) I added something to info.plist that asks for permission to use camera and thats all. 3) I incorporated this new code for images for creating a picture for CreateAStop, creating a cover image for CreateAHunt, and for taking a picture when completing a stop in CompleteAStop.
Everything with the new code seems great and works except the one thing is that when you "open" the camera or camera roll, the space it takes up on the screen is just the portion of the screen the icon takes up. I feel like that isn't too difficult to accomplish I just dont want to rn.

Also, for CompleteAStop, currently it doesn't take the coordinates of the picture I took, but I think I may know how to do that and I'll work on that.

That is all! Lmk the questions u have.

I am shaking in my boots for survivor!!!