got it to work (for the most part). maps are now visible for search bar and for a specific hunt. a couple things: at one point, there can be three tabbed views at once (which looks uggo so we prob should fix that), for some reason, when you "go to hunt" from the searching map, the tabbed view isn't there (not sure why), and finally, for the map for a specific stop, sometime you have to go back and forth a couple times for the pins to load (prob an issue somewhere in the code that isn't quite right in loading the locations of the pins).
for the map for hunts, i have the alert tell the user to complete the stop in the list. we can keep that or allow them to complete it on the map, both should be fine. also, the map starts out really zoomed out and there is not current location blue circle thingy (idk if that's possible/easy enough to do). the zoom should be easily fixable, idk about the other one.